many work done

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Posts;
+use App\Tags;
 use App\User;
 use Illuminate\Http\Request;
 use Intervention\Image\Facades\Image;
@@ -28,7 +29,9 @@ class PostsController extends Controller
      */
     public function create()
     {
-        return view('posts.create');
+        return view('posts.create', [
+            'tags' => Tags::all()
+        ]);
     }
 
 
@@ -38,21 +41,18 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-        $addPostData = $request->validate([
-            'post_title' => ['required', 'string'],
-            'post_content' => ['required'],
-            'post_thumbnail' => ['required', 'image'],
-        ]);
 
+        $addPostData = $this->validatePosts($request);
         $post_thumbnail_path = request('post_thumbnail')->store('uploads', 'public');
         $image = Image::make(public_path("storage/{$post_thumbnail_path}"))->fit('500', '500');
         $image->save();
 
-        auth()->user()->posts()->create([
+        $post = auth()->user()->posts()->create([
             'post_title' => $addPostData['post_title'],
             'post_content' => $addPostData['post_content'],
             'post_thumbnail' => $post_thumbnail_path,
         ]);
+        $post->tags()->attach(request('post_tags'));
         return redirect('/user-profile/' . auth()->user()->id);
     }
 
@@ -70,8 +70,17 @@ class PostsController extends Controller
     public function destroy($post)
     {
         $user_id = auth()->user()->id;
-        $delete_post_id = Posts::where('id', $post)->where('user_id', $user_id)->get();
-        $delete_post_id->delete();
+        $delete_post_id = Posts::where('id', $post)->where('user_id', $user_id)->delete();
         return redirect('/user-profile/' . auth()->user()->id);
+    }
+
+    public function validatePosts($request)
+    {
+        return $request->validate([
+            'post_title' => ['required', 'string'],
+            'post_content' => ['required'],
+            'post_thumbnail' => ['required', 'image'],
+            'post_tags' => ['exists:tags,id'],
+        ]);
     }
 }
